# back/src/db/schemas/comment.js
import { Schema, model } from "mongoose";

const CommentSchema = new Schema(
  {
    boardId: {
      type: Schema.Types.ObjectId,
      ref: "Board",
    },
    writer: {
      type: Schema.Types.ObjectId,
      ref: "User",
    },
    content: {
      type: String,
      required: true,
    },
  },
  { timestamps: true }
);

const Comment = model("Comment", CommentSchema);

export { Comment };