# back/src/modules/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"Swagger for Cocktail API","description":"Open API docs with Swagger","license":{"name":"MIT","url":"http://localhost:5000"}},"host":"localhost:5000","basePath":"/","tags":[{"name":"User","description":"User 로그인, 회원가입 API"},{"name":"Cocktail","description":"칵테일 조회 API"}],"schemes":["HTTP"],"consumes":["application/json"],"produces":["application/json"],"paths":{"/":{"get":{"tags":["Main"],"summary":"Main Page Get","responses":{"200":{"description":"OK"}}}},"/boardList":{"get":{"tags":["Board"],"summary":"게시판 전체 조회","responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/board":{"post":{"tags":["Board"],"summary":"게시판 생성","description":"이미지는 formData로, key=img","parameters":[{"in":"body","name":"body","title":"body","content":"body","schema":{"$ref":"#/definitions/Board"}},{"name":"authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/board/{id}":{"get":{"tags":["Board"],"summary":"특정 게시글 조회","parameters":[{"in":"path","name":"id","description":"게시글 oid"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}},"delete":{"tags":["Board"],"summary":"게시글 삭제","parameters":[{"name":"authorization","in":"header","description":"JWT access token","required":true,"type":"string"},{"in":"path","name":"id","description":"게시글 oid"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}},"put":{"tags":["Board"],"summary":"게시판 수정","parameters":[{"in":"body","name":"body","content":"body","title":"body","schema":{"$ref":"#/definitions/Board"}},{"name":"authorization","in":"header","description":"JWT access token","required":true,"type":"string"},{"in":"path","name":"id","description":"게시글 oid"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktails/{name}":{"get":{"tags":["Cocktail"],"summary":"특정 칵테일 조회","parameters":[{"in":"path","name":"name"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktails/{word}":{"post":{"tags":["Cocktail"],"summary":"특정 단어 포함 칵테일 조회","parameters":[{"in":"path","name":"word"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktails/page/{offset}":{"get":{"tags":["Cocktail"],"summary":"칵테일 전체 조회","parameters":[{"name":"offset","in":"path","type":"integer"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktailmbti":{"post":{"tags":["Cocktail"],"summary":"mbti 결과값 리턴","parameters":[{"name":"mbti","in":"body","description":"[w, x, y, z] => [(w > 0 = E: w < 0 = I), (w > 0 = N: w < 0 = S), (w > 0 = T: w < 0 = F), (w > 0 = J: w < 0 = P),]","required":true,"mbti":"body","schema":{"$ref":"#/definitions/mbti"}}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktails/rank":{"get":{"tags":["Cocktail"],"summary":"랭킹 1 ~ 10 칵테일 정보 조회","responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktail":{"post":{"tags":["Cocktail"],"summary":"칵테일 정보 입력","parameters":[{"in":"body","name":"body","ingredient":"body","imageUrl":"body","taste":"body","description":"body","userId":"body","method":"body"},{"name":"Authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktail/modify":{"post":{"tags":["Cocktail"],"summary":"칵테일 정보 수정","parameters":[{"in":"body","name":"body","originName":"body","ingredient":"body","imageUrl":"body","taste":"body","description":"body","method":"body"},{"name":"Authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktail/{name}":{"delete":{"tags":["Cocktail"],"summary":"칵테일 정보 삭제","parameters":[{"in":"path","name":"name"},{"name":"authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktails/like":{"get":{"tags":["Cocktail"],"summary":"칵테일 이름 + 좋아요 수 조회","responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/board/comment":{"post":{"tags":["Comment"],"summary":"댓글 작성 API","parameters":[{"name":"authorization","in":"header","description":"access token은 토큰 앞에 'Bearer ' 꼭 붙여서 request 해야합니다.","required":true,"type":"string"},{"in":"body","name":"body","boardId":"body","content":"body","schema":{"$ref":"#/definitions/Comment"}}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/board/comments/{boardId}":{"get":{"tags":["Comment"],"summary":"특정 게시판 댓글 목록 조회 API","parameters":[{"in":"path","name":"boardId"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/board/comment/{id}":{"put":{"tags":["Comment"],"summary":"특정 댓글 수정 API","parameters":[{"in":"path","name":"id","description":"댓글 _id를 넣어주세요."},{"name":"authorization","in":"header","description":"access token은 토큰 앞에 'Bearer ' 꼭 붙여서 request 해야합니다.","required":true,"type":"string"},{"in":"body","name":"body","content":"body","schema":{"$ref":"#/definitions/Comment2"}}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}},"delete":{"tags":["Comment"],"summary":"특정 댓글 삭제 API","parameters":[{"in":"path","name":"id","description":"댓글 _id를 넣어주세요."},{"name":"authorization","in":"header","description":"access token은 토큰 앞에 'Bearer ' 꼭 붙여서 request 해야합니다.","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/like/{id}":{"post":{"tags":["Like"],"summary":"좋아요 추가","parameters":[{"in":"path","name":"id"},{"name":"Authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}},"delete":{"tags":["Like"],"summary":"좋아요 삭제","parameters":[{"in":"path","name":"id"},{"name":"Authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/likeList":{"get":{"tags":["Like"],"summary":"좋아요 전체 리스트","responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/cocktailLike/{id}":{"get":{"tags":["Like"],"summary":"칵테일이 받은 좋아요 리스트","parameters":[{"in":"path","name":"id"},{"name":"Authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/userLike":{"get":{"tags":["Like"],"summary":"유저가 좋아요한 리스트 목록","parameters":[{"name":"authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/rank10":{"get":{"tags":["Rank"],"summary":"Rank 상위 10개 데이터 반환","responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/rank/{name}":{"get":{"tags":["Rank"],"summary":"Rank 내 특정 칵테일 정보 반환","parameters":[{"in":"path","name":"name"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/register":{"post":{"tags":["User"],"summary":"register user Post","parameters":[{"in":"body","name":"body","email":"body","password":"body","schema":{"$ref":"#/definitions/User"}}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/login/modify":{"put":{"tags":["User"],"summary":"회원 정보 수정","parameters":[{"name":"authorization","in":"header","description":"JWT access token은 앞에 'Bearer ' 꼭 붙여서 request 해야합니다.","required":true,"type":"string"},{"name":"body","in":"body","description":"이메일, 패스워드, 이름 모든 값을 요청하지 않아도 됩니다.","email":"body","password":"body","schema":{"$ref":"#/definitions/User"}}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/login":{"post":{"tags":["User"],"summary":"로그인","parameters":[{"name":"body","in":"body","email":"body","password":"body","required":true,"schema":{"$ref":"#/definitions/User2"}}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/auth/google":{"get":{"tags":["User"],"summary":"구글 login","description":"해당 path로 로그인 후엔 /login로 돌아갑니다.","responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/refresh":{"get":{"tags":["User"],"summary":["Access Token 재발급"],"description":"access token은 토큰 앞에 'Bearer ' 꼭 붙여서 request 해야합니다.","parameters":[{"name":"authorization","in":"header","description":"JWT access token","required":true,"type":"string"},{"name":"refresh","in":"header","description":"JWT refresh token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/user/current":{"get":{"tags":["User"],"summary":["토큰 유효성 검사, 토큰 속 userId 정보 반환"],"description":"access token은 토큰 앞에 'Bearer ' 꼭 붙여서 request 해야합니다. 토큰 속 user 정보를 반환합니다.","parameters":[{"name":"authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}},"/withdrawal":{"delete":{"tags":["User"],"summary":"회원탈퇴","description":"탈퇴 시 유저가 작성한 게시글, 댓글은 writer 필드값 => null. 좋아요는 유지","parameters":[{"name":"authorization","in":"header","description":"JWT access token","required":true,"type":"string"}],"responses":{"201":{"description":"OK"},"404":{"description":"Failed. Bad Request"}}}}},"definitions":{"Board":{"properties":{"title":{"type":"string"},"content":{"type":"string"}}},"Cocktail":{"properties":{"name":{"type":"string"},"ingredient":{"type":"[string]"},"imageUrl":{"type":"string"},"rank":{"type":"Schema.Types.ObjectId"}}},"mbti":{"properties":{"mbti":{"type":"[integer]"}}},"Comment":{"properties":{"boardId":{"type":"string"},"content":{"type":"string"}}},"Comment2":{"properties":{"content":{"type":"string"}}},"Like":{"properties":{"giveUserId":{"type":"string"},"getCocktailId":{"type":"string"},"name":{"type":"string"}}},"Rank":{"properties":{"bane":{"type":"string"},"visitors":{"type":"string"},"years":{"type":"string"},"rank":{"type":"string"}}},"User":{"properties":{"email":{"type":"string"},"password":{"type":"string"},"name":{"type":"string"}}},"User2":{"properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}